- Fixed 64bit pointer defreference in OpenGL VertexBuffer Lock() #4

+++ Projects/Axiom/Source/RenderSystems/OpenGL/GLHardwareVertexBuffer.cs
@@ -134,7 +134,7 @@ namespace Axiom.RenderSystems.OpenGL
 
             isLocked = true;
 
-            return new IntPtr( ptr.ToInt32() + offset );
+            return new IntPtr( ptr.ToInt64() + offset );
         }
 
         /// <summary>
